requirements.txt
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.4.4
beautifulsoup4==4.8.1
calmap==0.0.7
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
dj-database-url==0.5.0
Django==2.2.4
django-bootstrap4==1.0.1
django-cors-headers==3.2.0
django-heroku==0.3.1
django-restframework==0.0.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
gmplot==1.2.0
gunicorn==20.0.4
idna==2.8
importlib-metadata==0.23
isort==4.3.21
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
matplotlib==3.1.1
mccabe==0.6.1
more-itertools==7.2.0
npm==0.1.1
numpy==1.17.2
optional-django==0.1.0
packaging==19.2
pandas==0.25.1
pluggy==0.13.0
psycopg2==2.8.4
py==1.8.0
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.2
pytest==5.2.1
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.2
requests==2.22.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
wcwidth==0.1.7
whitenoise==5.0
wrapt==1.11.2
zipp==0.6.0
